Fixing and adding missing imports and anootations | Refactoring code in Product service | Adjusting database script to matches entity type

+++ services/product/src/main/java/com/ecommerce/product/product/ProductPurchaseRequest.java
@@ -7,5 +7,5 @@ public record ProductPurchaseRequest(
     Integer productId,
 
     @NotNull(message = "Quantity is mandatory")
-    double quantity
+    Integer quantity
 ) {}
